Fix for https://github.com/nens/grondwaterinbrabant-site/issues/13
If percentiles are added to a flot graph, the label of the line these
percentiles is changed so that it describes the percentiles too. E.g.,
a label that used to read

"Grondwaterstand [cmNAP]"

may now read

"Grondwaterstand [cmNAP] (met 10% - 90% percentiel, 25% - 75% percentiel)"

The reason that the percentile data doesn't get its own legend items
is that the percentiles and the main data use the exact same color,
with different opacities. Opacities aren't shown in the legend, so
that makes it impossible to distinguish between the different legend
items. Therefore this solution (changing the label of the data that
gets the percentiles) was chosen.